+++ lint-staged.config.cjs
@@ -1,4 +1,4 @@
 module.exports = {
   '*.ts?(x)': ['eslint --fix --quiet'],
-  '*.{ts?(x),cjs,json}': ['prettier --write'],
+  '*.{ts?(x),cjs,json,md}': ['prettier --write'],
 };