+++ test/genOpenapi.js
@@ -1,19 +1,6 @@
-const { generateService, patchSchema } = require('../dist/index');
+const { generateService } = require('../dist/index');
 
 generateService({
-  schemaPath: `${__dirname}/example-files/openapi.json`,
-  serversPath: './apis/pet',
-});
-
-generateService({
-  schemaPath: `${__dirname}/example-files/openapi-display-enum-label.json`,
-  serversPath: './apis/all',
-  isGenJsonSchemas: true,
-});
-
-generateService({
-  schemaPath: `${__dirname}/example-files/openapi-display-enum-label.json`,
-  serversPath: './apis/app',
-  apiPrefix: "'/api'",
-  allowedTags: ["app"]
+  schemaPath: `${__dirname}/example-files/openapi-schemas-enum-array.json`,
+  serversPath: './apis/schemas-enum-array',
 });